Ginkgo has profiling support
- Ginkgo now correctly honors -cpuprofile, -memprofile, -blockprofile, and -mutexprofile
- Ginkgo can merge profiles generated by running in parallel
- Ginkgo can aggregate profiles in -output-dir
- Ginkgo preserves the test binary when these flags are set

+++ ginkgo/internal/compile.go
@@ -52,7 +52,10 @@ func CompileSuite(suite TestSuite, goFlagsConfig config.GoFlagsConfigType) TestS
 	return suite
 }
 
-func Cleanup(suites ...TestSuite) {
+func Cleanup(goFlagsConfig config.GoFlagsConfigType, suites ...TestSuite) {
+	if goFlagsConfig.BinaryMustBePreserved() {
+		return
+	}
 	for _, suite := range suites {
 		if !suite.Precompiled {
 			os.Remove(suite.PathToCompiledTest)
